Made Many Changes. See Description.
Made a Button that would mute and unmute the volume for click sounds on the website.
The user can swipe on the main navigation bar vertically to open and close the main nav menu.
The side swipe and button options have been removed. They make the website too clunky in my opinion.
Added a plugin that makes a qrcode for a dev server for my phone.