Version Update 3.3.3 (#62)
* Updated Dependencies and Project Version Number.

The Share popup button on the bottom left on the footer is now smaller on smaller screens.
The Update Box and Copyright page better describe the version of the website the visitor is on at the moment.

* Made Many Changes. See Description.

Made a Button that would mute and unmute the volume for click sounds on the website.
The user can swipe on the main navigation bar vertically to open and close the main nav menu.
The side swipe and button options have been removed. They make the website too clunky in my opinion.
Added a plugin that makes a qrcode for a dev server for my phone.

* Updated tests to use playwright and Vue.js specific testing modules.

It seems to work even better now.

* restarted dependencies.

* Update ci.yml

* should provide proper fixes now.

* removed playwright and check deps again.

* animations now use pointer events & update box displays time over version now.

* Gave some functionality to the escape key.