Made many updates. See Description.
Updated the offset parameter for the "goToPageSection" function so that it functioned as a de-facto "scroll-padding-top" over a "scroll-padding-bottom". It also can be used by the "mountWebData" function.
The Start Section Dropdown has a new button that takes the user to its respective section in the contact page.
Redirects have been updated to now accommodate more wildcard routes.