Version Update 3.3.4 (#63)
* Updated dependencies and project version number.

* Made buttons that allows the user to copy the username for a social media account of mine.

This applies to both the contact page and the start section of the main page.

* Made many updates. See Description.

Updated the offset parameter for the "goToPageSection" function so that it functioned as a de-facto "scroll-padding-top" over a "scroll-padding-bottom". It also can be used by the "mountWebData" function.
The Start Section Dropdown has a new button that takes the user to its respective section in the contact page.
Redirects have been updated to now accommodate more wildcard routes.

* Made a popup that indicates whether the battery on the user's device is low.

This popup also accommodates for the Update box.