Made some more updates to Gamepad compatibility.
Put the gamepad cursor speed component back in. For the time being, it controls ALL FOUR of the gamepads, but that will be changed by a new component in this update or a future one.
I also made a new particles background for the Gamepad page. I'll focus on make a new version of that page really soon.