Version Update 3.4.0 (#64)
* Updated Deps and starting updating gamepad compatibility.

The gamepad store will be updated to now be able to handle multiple Gamepads instead of just one.

* Fully integrated multi-gamepad compatibillity on my website.

I had to temprarily disable setting the cursor speed as it needs more UI adjustments, but everything else works as intended.

* Fixing testing problem (maybe)

* trying a better fix

* Made some more updates to Gamepad compatibility.

Put the gamepad cursor speed component back in. For the time being, it controls ALL FOUR of the gamepads, but that will be changed by a new component in this update or a future one.
I also made a new particles background for the Gamepad page. I'll focus on make a new version of that page really soon.

* Updated Deps and made slight visual fixes.

* Updated the GitHub directions to have links.

Also updated the annotation handler to handle more types of events.

* Made changes to the social contact elements. See Description.

The Contact buttons on the home page now go directly to the contact page when using SHIFT + click.
The Contact Dropdown on the Home page now scales down on smaller screens.
The Social tabs on the Contact page now glow should the route  hash match their element id.
Made other small design changes.

* Made last changes.

I did not edit the gamepad page, but I can do that in future edits.