Made some updates to the invalid and redirect page. See Description.
The page now changes its document head title and description based on if it needs to redirect or not.
It also has a redirect background particle set that dynamically changes.
Also updated aws sdk dependencies and added links to the deploy script on how I use the script in GitHub Actions.